+++ src/bklms_downloader/__init__.py
@@ -1,3 +1,3 @@
 """BK-LMS Downloader."""
 
-__version__ = "1.1.2"
+__version__ = "1.1.3"